refactor(contractFactory): add abstract extender to extend functionality on contract object method funcs
